standardize state (session) management
The book was using reagent.utils's version of a "session" (a reagent atom that tracks state).
However, luminus seems to prefer to just define that atom on its own. Which is fine and dandy
until you realize it needs to be referenced in other namespaces. It looks like what people do is
simply define it in a separate file [0] and be able to reference it in core.cljs and specific components;
which is kind of exactly what the reagent.utils version does[1]. So I decided to simply bring it back,
which seems to work for other people following the tutorial and other tutorial[2][3]
Still not sure why luminus stopped using that lib[4], other than minimalism.

[0] https://github.com/yogthos/yuggoth/blob/master/src-cljs/yuggoth/session.cljs
[1] https://github.com/reagent-project/reagent-utils/commit/1db7ebae29e95eec4165a8e3d0092b914724fb70#diff-eaa14f1f9020e7d2d5ab51f17a5e7325
[2] https://github.com/janraasch/picture-gallery/blob/master/src/cljs/picture_gallery/core.cljs#L3
[3] http://yogthos.net/posts/2014-08-14-Routing-With-Secretary.html
[4] https://github.com/luminus-framework/luminus-template/commit/b664e4b931f5c0f3b201b9be2afcf6b8d2705738